+++ src/platform/platform.h
@@ -28,7 +28,7 @@ TYPES
 /*-------------------------------------
 Platform callback functions
 -------------------------------------*/
-typedef uint32_t (*platform_get_time_func)(platform_t* platform);
+typedef float (*platform_get_time_func)(platform_t* platform);
 
 typedef boolean (*platform_load_file_func)(const char* filename, boolean binary, long* out__size, void** out__buffer);
 typedef FILE* (*platform_open_file_func)(const char* filename, long* out__size);
@@ -45,7 +45,7 @@ struct platform_s
 	void* context;	/* Context pointer for platform-specific data */
 
 
-	platform_get_time_func		get_time;	/* gets the time (in ms) between the previous frame and the current frame */
+	platform_get_time_func		get_time;	/* gets the current timestamp (in ms) */
 
 	/*
 	Allocates a temporary buffer and loads the specified file into the buffer.
